backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
asyncpg
pydantic-settings
python-dotenv
email-validator
httpx
certifi
PyJWT[crypto]
svix
